+++ clickup/spaces.go
@@ -75,11 +75,11 @@ type Space struct {
 	Name     string `json:"name"`
 	Private  bool   `json:"private"`
 	Statuses []struct {
-		ID         string `json:"id"`
-		Status     string `json:"status"`
-		Type       string `json:"type"`
-		Orderindex int    `json:"orderindex"`
-		Color      string `json:"color"`
+		ID         string  `json:"id"`
+		Status     string  `json:"status"`
+		Type       string  `json:"type"`
+		Orderindex float64 `json:"orderindex"`
+		Color      string  `json:"color"`
 	} `json:"statuses"`
 	MultipleAssignees bool `json:"multiple_assignees"`
 	Features          struct {
